Cleanup volume_attachment in case of reset-state attach_status
We have the ability to use reset-state to get the volume back
to a usable state, but current reset-state only changes
status, attach_status and migration_status for volume 'table'
of target volume.
This means target volume entry in the 'volume_attachment' table
keeps remaining even after resetting attach_status.
When the volume will be attached again, multiple instances
will be shown in "Attached to" of "cinder list" unexpectedly.

This patch adds a cleanup for volume_attachment of target
volume when attach_status equals 'detached'.

Change-Id: Ibd403ececffe5d10eff68304ba82f55ed7d43975
Closes-Bug: #1568034