Huawei: Modify array LUNs naming rules
We use the first 8 chars in the uuid string and hash
value of uuid to name array LUNs, We can establish
the relationship between the volume uuid and the array
lun name.

There is no need to rename lun on array in unmanage and
unmanage_snapshot, So we remove the rename calls.

Change-Id: Id4cdfc3b73a4c97fa2c14aa2afb361e643ab6369
Closes-Bug: #1570205