EMC VMAX - Oversubscription support
When max_over_subscription_ratio is 2.0 the pool are oversubscribed
by a factor of 2, or 200% over subscribed.  The reserved_percentage
is the high water mark where by the physical remaining space cannot
be exceeded.  For example, if there is only 4% of physical space left
and the reserve percentage is 5, user would not be permitted to
provision a volume.  This is a safety mechanism to prevent a scenario
that a provisioning request failing due to insufficient raw space.

Change-Id: I55614d05701d461d4b0b85f6f57c49d98b014641
Implements: blueprint vmax-oversubscription