NetApp: Remove type checks from manage workflow
In a previous release, Cinder began routing manage API calls
through the scheduler to validate the suitability of the
specified pool for a volume being managed. The NetApp
drivers no longer need to do this check, which is part
of the storage service catalog code that we're replacing
during Newton.

Partially implements: blueprint replace-netapp-cdot-ssc-module

Change-Id: I567b2a95844a5be9291eac4b7d256dcb997e5626