NetApp: Add iterator handling to client library
The forthcoming SSC client routines need to iterate over
potentially large numbers of objects.  So this commit adds
a reusable method that transparently handles API pagination
for getter APIs.  This code is fully unit tested, and it has
already shipped in Manila, so it's a known quantity.

Partially implements: blueprint replace-netapp-cdot-ssc-module

Change-Id: I3e2c6c14fa54f93010c46a26324b08f0a120694f