3PAR: Remove metadata that tracks the instance id
HPE added some additional tracking of the instanced ID metadata on the
3PAR storage arrays back in Icehouse. When a volume was attached, the
instance ID was stored on the 3PAR and removed on detach.
Initially this this was over SSH and moved to REST calls due to
performance issues. Many customers still have issues with this unused
feature on upgrades and under heavy load.

The tracking of the instance ID is no longer required by HPE and has no
affect on any of the Cinder functionality. This unused feature will also
be a problem when multi-attach lands, so it's being removed by this
patch.

Change-Id: Ic181e6598f37a727b41c093e0c09abe229183d6f
Closes-Bug: #1572665