Mark the v1 API as deprecated
The Volume API spec says that the v1 should be marked as "DEPRECATED"
when querying the version[1]. However, when that API was deprecated in
dce0424d9, it was marked as "SUPPORTED", so there is a conflict between
the spec and the actual behavior. This patch updates the query to match
the documentation. This affects how keystoneauth discovers versions.

[1] http://developer.openstack.org/api-ref-blockstorage-v1.html#listVersionsv1

Change-Id: I0c073fdc3ed1c51662ce06c05911e50efbdbb3a4