Move OVO methods from CinderObject to Persistent
When Conditional Update mechanism was introduced `conditional_update`
method was added to `CinderObject` instead of `CinderPersistenObject`.

Because this method was added to `CinderObject` other methods were added
as well.

And so we ended up with methods that make no sense in the base OVO
class, as they are only relevant for objects that have persistence and
can be retrieved from the DB.

This patch moves these methods (`get_by_id`, `refresh`,
`_get_expected_attrs`, and `conditional_update`) to
`CinderPersistentObject` where they should have been all along.

Change-Id: I9d9b76f250386bb5d16d6b3d0d5ced5f84b12260