Move initiator_data access to helper methods
This changes the flow for how a driver would access the initiator_data
information. Previously it was passed in at initialize_connection
and then a update dictionary merged in with the return of that method.
We will now just have drivers call helper util methods that can get
and set the data.

There are a few benefits of doing this:

* Drivers can synchronize access to the data, if they need to. It won’t
force any locks are anything on someone who doesn’t care. Future
iterations will make this do a conditional update so drivers won’t need
to use locks around it either.
* No more optional params on initialize_connection, meaning all drivers
have the same signature again.
* Less junk in the manager.
* Less db calls for everyone!

Change-Id: I9b90fafdbf290c7d7141b6ce4f83711f0fcd528f
Closes-Bug: #1578801
Closes-Bug: #1578864