Specify a user-agent in Pure volume drivers
Previously we were just using the default user-agent that the python
requests module would send. With newer versions of the pure storage
python module you can now specify a custom one.

This will help to identify on the backend what requests are being
sent by the VolumeDriver vs other python scripts running against the
same array.

As part of this we will also remove support for the older version of
the python module. The newer versions are released and easily upgraded
to.

Change-Id: Icf8595843e38381c6704fd6a9c66b752f91337a5
Closes-Bug: #1579194