Reorder DB API require_context decorators
In some cases, the context is checked after doing other
operations.

As this is something enforcing a security boundary,
we really want to check context before anything else,
and fail immediately, to mitigate anything that could
go wrong in the other decorators.  So, ensure the
context checks run first.

This also removes an extra volume_get() for
volume_admin_metadata_update() which was loading the volume
twice.

Change-Id: Ia6432de766b22e80e337b67a994cdb921ea14cc6