Differentiate thick and thin provisioning
If a driver reports thin_provisioning_support to True and
thick_provisioning_support to True on a pool that supports
both thin and thick volumes, and the user wants to create a
thick volume, the current logic in the scheduler would use
the wrong logic which makes decisions based on
provisioned_capacity_gb and max_over_subscription_ratio.

This patch tries to address the problem.

Implements: blueprint differentiate-thick-thin-in-scheduler
Closes-Bug: #1580866
Change-Id: I8b43624b934fadf1036fafa1f5e2b94b5317f81c