Unexpected function is called in _migrate_volume_generic
When calling _migrate_volume_generic with the volume created by
create_volume_from_snapsot or create_cloned_volume,
create_volume is not called to create dest-volume for migration.

Instead of the function, create_volume_from_snapshot or create_cloned_volume
is called to create the dest-volume.

This is because the source volume for migration has info of 'snapshot_id'
or 'source_volid'. So, This patch temporarily removes the keys and sets after
migration.

Change-Id: Ibeeb119b72c0a686000f175187c1bfa359bbf6a2
Closes-Bug: #1580013