Create encrypted volumes from images
Currently, when creating an encrypted volume from an image, Cinder
writes raw data to the encrypted volume.
As a result, Nova can't use these volumes.

This patch is to implement following function:
When creating an encrypted volume from an image, it encrypts the
data and writes to the volume.

This patch adds a new interface copy_image_to_encrypted_volume in
driver's API, vendor company can implement it by itself.

Change-Id: I213459193550198c570615e381797db5e08e2cce
Implements: blueprint improve-encrypted-volume
Closes-bug: #1482464
Closes-bug: #1465656