gpfs: remove a race in deleting volumes
"find" command may end with error if any of files in a target directory
is removed during its run. As we can/should ignore it and continue the
operation, we run "find" with -ignore_readdir_race option to stop
considering it an error.

Also, "find" runs with -print0 and -quit as it is a better practice.

Closes-bug: 1581670
Change-Id: Id26cb59d88082971465c8fcf0c150c858ab3bced