Fix backup using temp snapshot code path
- Non-disruptive backup using a temp snapshot works in Liberty
but was broken in Mitaka.
- Backup a snapshot directly without creating a temp volume
worked when the feature was first added in Mitaka but was broken
later in Mitaka.

This patch provides a fix as follows:
1. It checks an existing config option backup_use_same_host.
   By default, this option is False.
2. When the backup service starts, it checks the above option. If
   the option is True, the backup service will find volume manager
   on the current node and get volume backends.
3. If the option is True and backup_use_temp_snapshot returns True,
   volume service returns the snapshot to the backup service in
   get_backup_device and backup will be performed using the snapshot.
   Otherwise, the volume will be returned as the backup device and
   the backup will be performed using the volume.

This fix is a Mitaka backport candidate. After this is merged,
we will provide a more complete fix which allows backup using
temp snapshot to happen on a remote node as well and we will
also clean up the code to get volume backends on the backup node.

The unit test test_backup_volume_inuse_temp_snapshot in test_volume.py
is removed. This test was testing backup_volume in
cinder/volume/driver.py, but this code path is not used any more.
Backup now starts in create_backup in backup/manager.py which calls
_run_backup which calls _attach_snapshot in volume/driver.py. The new
test test_create_backup_with_temp_snapshot in test_backup.py tests
the new code path.

Note: The return value is changed for _attach_snapshot() in
volume/driver.py. It now returns a dictionary of attach_info.

Closes-Bug: #1578034
Closes-Bug: #1575886
(cherry picked from commit 763dd8cdc5c48b7771d081dcd8b9e37a8f6c0a71)

Conflicts:
	cinder/backup/manager.py
	cinder/tests/unit/test_volume.py

Change-Id: I2e0c115e1dacf9eea73803cdbb1452bfeb56d87c