ScaleIO actual volume size model update
Report back and update the actual volume size that is created in ScaleIO.

ScaleIO volumes support sizes that are multiplies of 8 GB.
This change updates the db with the real volume size.
However, it is a partial fix, because the quota is not updated.

The boolean property sio_round_volume_capacity specifies whether the
user allows ScaleIO to create volumes with a size that is
not a multiply of 8GB, hence causing ScaleIO to round it up.
When set to False, creation of volumes with a size that is different
than 8GB will fail.

Change-Id: I10e699eab98715cf7c8cd79533fc9279e81dfbf6
DocImpact:
Partial-Bug: #1561560
(cherry picked from commit 79d2d63ad868b6898bd13327f6b81712348883b6)