IBM Flashsystem: Add missing items to common
There are a handful of methods that are
intended to be inherited from this base class
that aren't defined, and some member variables
which are not guaranteed to be defined before use.

Add missing items to the base common class
to clean things up.

This removes seven issues from our pylint run.

Change-Id: I2a93852bbd5fc528f6f73d99212ac6040eaa004a