Improve DB volume check in LVM manage volume
LVM driver makes a call to volume.utils.check_alredy_managed_volume when
managing a new volume in method manage_existing, and this check method
is trying to retrieve the volume for the check, which means getting all
the volume information with joined tables and then converting it to an
ORM object.

This patch improves this check_already_managed_volume by using a simpler
EXISTS DB query.

Change-Id: I132b6bdd05524c955fcd404ed45624e1b1e98320