Implement OVO class init mechanism
This patch adds a mechanism to perform class initialization in Versioned
Objects.

One case where this is useful is when initializing the model class
attribute for all Persistent Versioned Objects.

We are currently trying to initialize the model class attribute for all
Versioned Objects in the registration hook, which is wrong on two
fronts, one is the place -it should be done in the
CinderPersistentObject- and the other is that we are doing it for all
classes -this could be easily fixed by checking if it's a
CinderPersistentObject class-.

This patch adds method `cinder_ovo_cls_init` to CinderPersistentObject
and changes the registration hook to call this method on any OVO class
that is registered if it exists.

Change-Id: I44c0fb41e3a2e6d615dc87ee6fe32d6a3ecf102a