Fix attachment_get_by_* to return entire list
Commit 10d54216871fde27172fbcb1a3c9bfec59b824c2 added the
ability to query volume_attachments by host and instance_uuid.

Minor bug however is that the original implementation also used
a "first()" filter, so regardless of how many attachments might
exist, only one was being returned.

The volume_attachments object code assumes we receive a list, as
do a number of other places.  So, this patch changes the names of
those get methods to "get_all_by_xxx" to eliminate any confusion
and updates things to use the list appropriately.

Change-Id: I63e4faf8a7ee3528a4dfbd7bf4e61b607ad86841
Closes-Bug: #1585690