Remove IN_USE from ConsistencyGroupStatus enum
It seems that we've added IN_USE as a valid CG status, which actually
isn't valid. It is used only in tests. This commit removes it from the
enum and replaces occurrences in tests with equivalent valid statuses.

As CG should not ever get that status in runtime, no new version of the
object or compatibility routine is added.

TrivialFix

Change-Id: Id8290e73a38a941a2a84391ded6ddbcb87be411e