Fix size exceed of dd action when clearing a volume
This patch closes a bug that may result in "No space left" error
of dd command.

If user set the config item volume_clear_size to larger value than
a size of a volume, the "dd" command will get "No space left"
error when deleting the volume.
This change rejects volume_clear_size settings that are larger
than 1GiB, because there isn't really a reason to set that.
That would be safer all-around.

DocImpact

Closes-Bug: #1588310
Change-Id: I690d62bc728c85c9795a87f84b52f1a14474933d