Huawei: Support iSCSI configuration in replication feature
In most case, users need much more features to attach volume to VM,
such as iSCSI multipath, CHAP, etc. In Huawei cinder driver replication
feature, we need to fit the case when using iSCSI driver.

We add a new key word into replication_device, called iscsi_info. The
detail as follows:

    replication_device = backend_id:huawei-replica-1,
        storage_pool:pool_1,
        san_address:san_url_1;san_url_2,
        iscsi_default_target_ip:192.168.1.100,
        san_user:admin,san_password:passwd,
        iscsi_info: {Name:xxx;CHAPinfo:user#password;ALUA:1;
                     TargetIP:0.0.0.0;TargetPortGroup:xxx};
                    {Name:xxx;CHAPinfo:user#password;ALUA:1;
                     TargetIP:0.0.0.0;TargetPortGroup:xxx}

All key words in iscsi_info are case insensitive. The meaning of them
are the same with that in huawei_config.xml, which always used in
master array.

Change-Id: I4a0cce1858c3a8539a62e7dae6e1092d4e86707f
Closes-Bug: #1588736