SolidFire driver should return ceiling of volume size
The function manage_existing_get_size returns the floor of volume size
in solidfire driver. But in the driver API:"When calculating the size,
round up to the next GB." The solidFire driver should be consistent
with the driver API. This patch fixes it by returning the ceiling of
volume size.

Change-Id: Icc4f893b9e18dc525b85b45f3307663b6d8884ed
Closes-Bug: #1586306