Capability lists in Cinder scheduler
Some volume drivers would like to report pool capabilities such as
'compression': [True, False] to denote a pool which can simultaneously
handle volumes with compression enabled and others with compression
disabled. This commit enhances the CapabilitiesFilter to handle this
use case. An extra spec value will be compared to a capability list
and will be passed by the filter if it matches anything in the list.

This feature shipped in Manila during Mitaka, and this commit is
line-for-line identical to that one. 100% unit test coverage is
included.

Change-Id: I0d70360c5c3e7824d226de3b2eb5cfdfb94621c2
Implements: blueprint support-scheduler-capability-lists