Add generic code for online data migrations
To limit downtime during upgrade we've decided to ban data migrations
done in migrations scripts and do them online during the runtime of the
cloud. This commit implements generic infrastructure for that.

This was part of a commit adding a mapping table between volume types
and consistency groups, but I've separated it because it's needed
independently. Also the aforementioned commit will probably be obsolete
when we finish generic volume groups efforts.

UpgradeImpact
Related: blueprint online-schema-upgrades
Change-Id: I655d47c39e44251cbbad6e4b3fbc8e8d73995d30