add FIXME tag for extra argument 'topic' in scheduler rpc api
Some functions in scheduler rpc api receive unused argument
'topic'. We can't fix it now as we may interact with old services.
It can only be fixed when we do a major version bump and drop the
old compatibility. Now we add two FIXME tags in case we forget it
one day.

Change-Id: Iafd60bb51a2d7c8eab909d35f4b4de120a1172bb
Partial-Bug: #1593149