Added update-host for CGs in cinder-manage
Currently in cinder-manage, "update_host" command for consistency
groups is missing.

Hence, in a similar way like volumes and backups, 'update_host'
is added for Consistency Groups as well, because if host
configuration is changed then even if we change volumes and
backups using cinder-manage, we would have to go to the DB to
change our CGs.

Change-Id: I820ae9d17460a5aa7b3b2e41afc861e8bc0265a3
Closes-Bug: 1591153