VMware: Honor vmware_cluster_name during retype
The vCenter clusters the VMDK driver should use for volume placement
are specified using the config option 'vmware_cluster_name'. During
retype, the VMDK driver may relocate the volume from one vCenter
cluster to another. But, the 'vmware_cluster_name' option is ignored
during retype and hence the volume may be relocated to a vCenter
cluster not configured for the VMDK driver. This patch fixes the
problem by using a utility function which considers the option
'vmware_cluster_name'.

Change-Id: I9d6889ca3095de49f27959fdf63b274de4467b79
Closes-bug: #1594759