VMware: Fix create vol from streamOptimized image
Change I247bf2a072e195c3918f222a9770f8018690db16 renamed a keyword
argument in get_create_spec. The _fetch_stream_optimized_image
method which creates a volume from streamOptimized image uses this
method, but it was not updated to use the new keyword argument. The
VMware CI couldn't catch this because we use preallocated image for
tempest tests. This patch updates the keyword argument to fix the
issue.

Change-Id: I20cbc8f137fe31302770cc0b2f5dde1f23339191
Closes-bug: #1596907