3par driver handles concurrent host create conflict
One of the step in creating a FC VLUN is to create a host on array.
When there happens to have simultaneous request for creating same host,
hpe3parclient will create host for one and return HTTP 409 conflict for others,
with error code 73, message "host WWN/iSCSI name already used by another host".
This was not handled by 3PAR driver and the HTTP exception would propagate up to
volume manager where it would log it. This patch addresses this bug by catching
this exception and reusing the same host.

Change-Id: I66e4f80696c782ab869ef0d7848fc6f3b3100c04
Closes-Bug: #1597454