Remove translation from config options
There were a few cases where config option help text was marked
for translation. This text is just used in the generated config
sample and does not ever actually get translated for end user
runtime display. This results in extra strings for the translation
teams to translate with no actual benefit.

Change-Id: I8594533d0e3a61da00bd42ac0f7d8e3da75616fc