Don't use shred for volume clearing
Volume clearing is intended to provide safety against
data leaking between different volumes/tenants when
backing storage is reused.  dd is sufficient to
accomplish this, and since 'shred' does not add anything
meaningful and has not seen much real-world use,
we should remove it for simplicity.

This change makes 'volume_clear=shred' call dd and
warns that this option will be removed.

Change-Id: I190dc22c5edb6efc56b98c5eac870a35c03fcd3f