Remove generate_driver_list dependency on cur dir
tools/generate_driver_list.py assumes that the current directory is
the Cinder root directory. If the script is run from a different
directory, it doesn't generate the driver info. This patch changes
the current directory to the root directory before calling the
utility functions to generate the driver list.

Closes-bug: #1604269
Change-Id: I35829003868531ccd76ee26e83e0b51067984a6a