Update OVO instance on destroy method call
None of the Versioned Objects destroy methods we have update the
instance fields, so after doing a destroy on an instance the status,
deleted, and deleted_at fields will still be the same as they were
before, leaving the instance in an inconsistent state.

This is important because our workers decorator for the HA A-A cleanup
mechanism relies on the status of the resources being properly updated
by the operations.

Change-Id: I4f15356931007e2ca725e477327c1ef67c7db487
Closes-Bug: #1604347