Use original volume OVO instance in create flow
Current manager flow for creating a volume uses the volume's id as input
instead of the volume OVO instance that we already have.

This patch changes this and uses the volume OVO for the whole flow which
has the benefit of having a consistent/updated instance once the flow
has finished.

Having the original instance updated is important because our workers
decorator for the HA A-A cleanup mechanism relies on the status of the
resources so they need to be properly updated.  Before this patch the
status from the volume OVO would remain in 'creating' regardless of the
result of the operation.

Change-Id: I7f45a151675ad976ca96f4de0cc34f19a31912f2