NetApp: Report hybrid aggregates in volume stats
NetApp cDOT controllers can mix SSDs and spinning disks
in the same aggregate, where the SSDs are used for a
cache.  This commit reports the hybrid aggregate attribute
to the scheduler for each pool for use in extra specs
matching.  This commit also ensures that all aggregate
values reported to the Cinder scheduler correctly include
a vendor-specific prefix.

Implements: blueprint netapp-report-cdot-hybrid-aggrs
Change-Id: I8fbfbe3ffaad5fe89db03d2a4e567e9b35e59d5b