Fail manage operations if service is down
Existing code will fail at API level if a service is disabled and we try
to manage/get_manageable volumes/snapshots on that service, but it won't
fail at API level if the service is down, and the message will wait in
the message queue until the service is up again, which is misleading.

This patch changes this behavior and makes manage_existing,
manage_existing_snapshot, get_manageable_volumes, and
get_manageable_snapshots, fail at the API if the service is down.

Change-Id: Ifd56a46fce9ac2fe72a30820c05a5cf9adee229f
Closes-Bug: #1605109