VMAX driver - Cleanup of Initiator group fails
When the last volume has been deleted from the Storage
Group and the Masking View has been deleted and the
initiator group is no longer in use, the deletion of
the Initiator group fails. The initiator group is actually
deleted off the array, but the code is throwing up an
error when it uses the now deleted initiator group
name to search for remaining masking views. The
solution is to check if the initiator group has been
deleted by common.terminate_connection before searching
for remaining masking views.

Change-Id: I79559789459f73c4adebf010d84cbeedd25f7f74
Closes-Bug: #1605193