Fix volume upload-to-image for vhd disk-format
It was previously assumed that glance did not support qed
disk-format.
qemu-img still uses the legacy 'vpc' format name as a
parameter. When executing 'qemu-img convert', vhd was
being passed as a parameter which is invalid.

This fix adds 'vhd' to the valid disk formats and passes
'vpc' to 'qemu-img convert', which is correct use of the
command. Also added 'vhdx' to the valid disk formats.

Change-Id: I24e6b78505a2c77b309c79ef11466c08bad1c50a
Closes-Bug: #1585612