Merge two fake drivers into the one FakeLoggingVolumeDriver
Two different fake drivers for unit-tests look confusing. We are able to
test anything we want with only one fake driver.

Current LoggingVolumeDriver doesn't actually log anything. It just
stores all drivers actions in a list.

Now, we can add real logging to FakeLoggingVolumeDriver and use it both
for unit and functional tests.

It's a first patch in a series for unit tests refactoring. More manager
and LVM driver tests will be proposed in a following patches.

Change-Id: I383bcdb531c7d52c0fdbb6875de73f1274a92854