Pin RPC server's serializer to min obj version
We're passing object version pin on CinderObjectSerializer
initialization in RPC clients (rpcapis), but not in RPC servers
(managers). This means that objects returned from RPC calls aren't
backported and older service can get a response with an object version
it doesn't understand. As on RPC server side we don't know who called
us, we need to consider minimum object version among all the binaries,
not only the one we're running.

This commit implements that behavior.

Closes-Bug: 1609304
Change-Id: Ibc708eb6ca6d792bec0f38abb4e1b5f58488a414