Prevent adding same OVO version twice to history
Some patches go through a lot of rebases and changes and may end up
mistakenly reusing the same history version as a prior entry.

This is usually caught on review, but just to be sure that this never
gets merged this patch adds a check to the CinderObjectVersionsHistory
add method.

Change-Id: Id5ac41ffcce86c3499e49ad60ac46038567e9fdb