Use OVOs to set errors in manage volume/snapshot
Currently when we are managing volumes and snapshots and we run into an
error that requires changing the status to error we don't make use of
the OVO instances we already have and instead pass just the id to a
helper function that in some cases retrieves the resource again fro the
DB before doing the update.

This patch leverages the OVO instance we already have to consolidate
both helper methods -error_out_volume and error_out_snapshot- into using
error_out method.

Doing this not only simplifies the code but it also removes one extra DB
query.

Change-Id: I8d3b47dd71ee616bf00ba98ad1da0935356ebe74