Add get_manageable_* methods to Pure drivers
This implements basic support for the new driver methods
get_manageable_volumes and get_manageable_snapshots. It
uses the "client side" sorting/pagination like the
reference driver, and to allow support for largest Purity
REST API version range. Future additions can do some of
the sorting and stuff via the REST API on the FlashArray.

We only mark volumes as unsafe if they are already managed
or connected to hosts. For snapshots it is only if they
are already managed by cinder.

Change-Id: I91d808021995921fa51a9f99e3a1e9c7091af865