Refactor cinder.utils._copy_volume_with_path
Our current _copy_volume_with_path method builds the command it wants to
execute in a non ordered way, meaning that it first builds the end of
the command to execute, then the middle, then the second part, and then
the first part, which makes it hard to follow.

It's not only hard to follow but inefficient, since we keep creating
unnecessary lists.

This patch does a little refactor of the method so that it is built from
the start of the command (with the prefix) to the end.

Change-Id: I88f941d6f9b3be59b849e4d894eeb06dbbd0d103