Fix exceed of dd action when copying with path
Our _copy_volume_with_path uses the blocksize passed to it as the unit
to copy data and then tells dd to copy N times this value based on the
size in MB.

The problem arises when the size of the volume is not a multiple of the
blocksize -ie: deleting a 1GB volume with a 3MB blocksize- because we
try to block more than we can because we do a ceiling on the
calculation.

One case where this can be seen is when clearing a volume on deletion,
but that's not the only case, as other drivers use the copy_volume from
the utils file.

Instead of changing the way we do the calculation and prevent the last
partial block from being zeroed -which may be valid for clearing a
volume but won't be for copying a volume since some data won't get
copied- what this patch does is change the unit of measurement for dd,
so instead of counting by blocks we will be counting by bytes and let dd
do the smart thing and just partially copy the last block.

Change-Id: Icdbf32b08fe641c5e32823eb8593e326d0171ca6
Closes-Bug: #1591031