Add tempest tests for Consistency Groups
These tempest tests add coverage for all the API's related to
consistency groups in Cinder. They were originally proposed
for upstream tempest in https://review.openstack.org/#/c/252213/
but were not a good fit since they aren't supported by the
reference driver. They have been modified to work as part
of the in-tree tempest plugins for Cinder now.

The tests are behind a new config option for tempest, which
in turn is part of a new config group called 'cinder'. This
was added to avoid any collisions with the 'volume-features-enabled'
or 'volume' groups already in the upstream tempest tests.

To enable them set the following in tempest.conf

[cinder]
consistency_group = True

Then make sure to run tempest with the 'all-plugin' tox environment.

Don't forget to update policy.json to allow for CG API's to be called..

Change-Id: I772ea13ca156e71620d722eee476f222a8653831
Co-Authored-By: Xing Yang <xing.yang@emc.com>