Fix mysql binary comparison
MySQL doesn't do case sensitive filtering by default, so our DB API
forces comparisons to be case sensitive when looking for hosts.

Existing code was not being tested at the gate because our check to
confirm a MySQL connection was too strict and the actual operator used
for the comparison was wrong.

This patch improves the connection type check to be more inclusive -now
the code path it will be run at the gate- and it fixes the operator to
use the right one.

Change-Id: Ie2da2eabbab1eb64106f99845dc82a60de14f036
Closes-Bug: #1613624