Performance issue for VNX attache LUN
The VNX driver experience some performance issue during initialize
connection.

After some analysis, we root caused the problem to two issues.
1. Storage group instance was retrieved every time it's needed.  Extra
CLI calls were made to update the host LUN information in the storage
group.  These CLI calls could be saved if we cache the storage group
instance in the client.
2. Storage group instance was force updated even though we didn't do
any configuration update for the initiator port map.  This is an
implementation error.  Use `any` to check whether the list of updated
initiator port map is empty or not and do not call force update if it's
empty.

Change-Id: I6f101a73d07cb58859ff25aad9411f18b67aa67f
Co-Authored-By: Ryan Liang <ryan.liang@emc.com>
Closes-Bug: #1615621