VMAX driver - Removal of iscsiadm from vmax cinder
Currently the controller node needs a masking view
with all ports exposed. This is unnecessary overhead
as the iscsiadm command does not need to be run on
the controller, it just needs to be run on the
relevant compute node. Fix ensures correct information
is presented to the compute node, without first doing
the expensive iscsiadm check.

Change-Id: Ia6e0b15f0f5696d4af02bad11c7818fb53cc6f90
Closes-Bug: #1587969