Recover volume status after remove_export
Currently Cinder recover volume's status before remove_export,
which leads that recreated iSCSI targets may be deleted and
attachment fails.

Let's consider the case: A volume is attached to instance A, and then
detached. During detach, Cinder sets volume's status to available and
remove iSCSI targets.
At the exact time the volume is reset to
available but targets are not removed, the volume is attached to
instance 2 and targets are created again. And then former remove_export
removed the recreated targets, and then it leads attachment fail.

This patch is to move remove_export after changing volume's
status.

Change-Id: I914d4badd9b93e9dc8f4ebc4f74314afe95bf855
Closes-Bug: #1616719